+++ strato-vscode/src/extension.ts
@@ -6,7 +6,6 @@ import * as path from 'path';
 import { ContractsProvider } from './contracts';
 import { CirrusProvider } from './cirrus';
 import { NodesProvider } from './nodes';
-import { ProjectActionProvider } from './project';
 import { activateStratoDebug } from './activateStratoDebug';
 import { subscribeToDocumentChanges } from './diagnostics';
 import { rest, importer } from 'blockapps-rest';
@@ -23,209 +22,6 @@ export async function activate(context: vscode.ExtensionContext) {
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
-
-	// Imports the project directory into the VS Code workspace context
-	context.subscriptions.push(vscode.commands.registerCommand('strato-vscode.importProject', async () => {
-		// Import project directory into workspace
-		const options: vscode.OpenDialogOptions = {
-			title: 'Import project directory',
-			openLabel: 'Import project',
-			canSelectMany: false,
-			canSelectFiles: false,
-			canSelectFolders: true
-		};
-
-		const folderUri = await vscode.window.showOpenDialog(options);
-		if (folderUri && folderUri[0]) {
-			console.debug(`importProject/Selected folder: ${folderUri[0].fsPath}`)
-
-			const projectName = folderUri[0].fsPath
-			const workspaceFolderUri = vscode.Uri.parse(projectName);
-			const numFolders = (vscode.workspace.workspaceFolders || []).length;
-			vscode.workspace.updateWorkspaceFolders(0, numFolders, { uri: workspaceFolderUri });
-
-			context.workspaceState.update('strato-vscode.workspaceDir', projectName)
-				.then(() => console.debug(`importProject/set strato-vscode.workspaceDir to ${projectName}`))
-			vscode.window.showInformationMessage(`STRATO project succesfully imported to workspace at ${projectName}`)
-		}
-	}))
-
-
-	// Builds and installs project dependencies
-	context.subscriptions.push(vscode.commands.registerCommand('strato-vscode.buildProject', async () => {
-		vscode.window.showInformationMessage("Building your dApp...")
-
-		// Show the directory paths in a quickPick dialog
-		const folderPath: string = vscode.workspace.workspaceFolders ? vscode.workspace.workspaceFolders[0].uri.fsPath : ""
-		console.debug(`buildProject/folderPath: ${folderPath}`)
-		const subDirectories = getSubdirectories(folderPath).map(a => path.basename(a)).filter(b => !b.startsWith('.'))  // Ignore .git, .vscode, etc
-		console.debug(`buildProject/subDirectories: ${subDirectories}`)
-
-		// Have the user select the directories where the build scripts
-		// are located then install the project dependencies
-		let backendDir, frontendDir
-		vscode.window.showQuickPick(subDirectories, { placeHolder: 'Select the backend directory', ignoreFocusOut: true })
-			.then(sp => {
-				if (sp) {
-					backendDir = sp
-					context.workspaceState.update('strato-vscode.backendDir', backendDir)
-						.then(() => console.debug(`buildProject/backendDir: ${backendDir}`))
-					runCommand(`cd ${folderPath}/${backendDir}; yarn install; cd ..`)
-				}
-			})
-			.then(() => {
-				vscode.window.showQuickPick(subDirectories, { placeHolder: 'Select the frontend directory', ignoreFocusOut: true })
-					.then(sp => {
-						if (sp) {
-							frontendDir = sp
-							context.workspaceState.update('strato-vscode.frontendDir', frontendDir)
-								.then(() => console.debug(`buildProject/frontendDir: ${frontendDir}`))
-							runCommand(`cd ${folderPath}/${frontendDir}; yarn install; cd ..`)
-						}
-					})
-			})
-			.then(() => vscode.window.showInformationMessage("Building your dApp..."))
-	}))
-
-
-	// Deploys the smart contracts to the targetted node using selected deploy scripts
-	context.subscriptions.push(vscode.commands.registerCommand('strato-vscode.deployProject', async () => {
-		// Check if .env exists
-		const folders = vscode.workspace.workspaceFolders || [];
-		const cwd = folders[0].uri.path
-		const backendDir = context.workspaceState.get('strato-vscode.backendDir') || 'backend'
-		console.debug(`deployProject/backendDir: ${backendDir}`)
-
-		const envPath = `${cwd}/${backendDir}/.env`
-		if (!fs.existsSync(envPath)) {
-			vscode.window.showWarningMessage('.env file does not exist, creating one for you. Consult your project README for additional help.')
-			runCommand(`touch ${envPath}`)
-			return
-		}
-
-		// Runs 'CONFIG=mercata yarn deploy' using backend/package.json
-		// TODO(moncayo): should enumerate all the test:<test_case> commands in package.json
-		const cmd: string = vscode.workspace.getConfiguration().get('strato-vscode.deployProjectCommand') || '';
-		runCommand(`cd ${cwd}/${backendDir}`)
-		runCommand(cmd)
-		runCommand('cd ..')
-
-		vscode.window.showInformationMessage("Deploying your dApp...")
-	}));
-
-
-	// Runs the project backend server
-	context.subscriptions.push(vscode.commands.registerCommand('strato-vscode.runServer', async () => {
-		const backendDir = context.workspaceState.get('strato-vscode.backendDir') || 'backend'
-		const cmd: string = vscode.workspace.getConfiguration().get('strato-vscode.runServerCommand') || '';
-		runCommand(`cd ${backendDir}`)
-		runCommand(cmd)
-		runCommand('cd ..')
-	}));
-
-
-	// Runs the project tests against the server
-	context.subscriptions.push(vscode.commands.registerCommand('strato-vscode.testServer', async () => {
-		const backendDir = context.workspaceState.get('strato-vscode.backendDir') || 'backend'
-		const cmd: string = vscode.workspace.getConfiguration().get('strato-vscode.testServerCommand') || '';
-		runCommand(`cd ${backendDir}`)
-		runCommand(cmd)
-		runCommand('cd ..')
-	}));
-
-
-	// Runs the project UI
-	context.subscriptions.push(vscode.commands.registerCommand('strato-vscode.runUI', async () => {
-		const frontendDir = context.workspaceState.get('strato-vscode.frontendDir') || 'ui'
-		const cmd: string = vscode.workspace.getConfiguration().get('strato-vscode.runUICommand') || '';
-		runCommand(`cd ${frontendDir}`)
-		runCommand(cmd)
-		runCommand('cd ..')
-	}));
-
-
-	// Runs the project tests against the UI
-	context.subscriptions.push(vscode.commands.registerCommand('strato-vscode.testUI', async () => {
-		const frontendDir = context.workspaceState.get('strato-vscode.frontendDir') || 'ui'
-		const cmd: string = vscode.workspace.getConfiguration().get('strato-vscode.testUICommand') || '';
-		runCommand(`cd ${frontendDir}`)
-		runCommand(cmd)
-		runCommand('cd ..')
-	}));
-
-
-	// Creates a private chain on the targetted node
-	// vscode.commands.registerCommand('contracts.createChain', async (element) => {
-	// 	const { nodeId } = element;
-	// 	const user = await getApplicationUser(nodeId);
-	// 	const config = getConfig() || {};
-	// 	const nodeOptions = { config, node: nodeId };
-	// 	const userAddress = await rest.getKey(user, nodeOptions);
-	// 	if (vscode.window.activeTextEditor) {
-	// 		const doc = vscode.window.activeTextEditor.document;
-	// 		if (doc.uri.path.slice(-4) === '.sol') {
-	// 			let src: any = doc.getText();
-	// 			let srcMap = {}
-	// 			const folders = vscode.workspace.workspaceFolders || [];
-	// 			if (folders.length > 0) {
-	// 				const serverPath: string = vscode.workspace.getConfiguration().get('strato-vscode.serverPath') || '';
-	// 				const currentFolder = folders[0]
-	// 				const folder = currentFolder.uri.path;
-	// 				// eslint-disable-next-line import/no-mutable-exports
-	// 				const dirPath = `${folder}/${serverPath}`
-	// 				srcMap = await importer.combine(doc.uri.path, true, dirPath);
-	// 				srcMap[importer.getShortName(doc.uri.path)] = doc.getText();
-	// 				src = Object.values(srcMap).reduce((str, fileContents) => `${str}\n${fileContents}`, '');
-	// 			}
-	// 			try {
-	// 				const chainLabel = await vscode.window.showInputBox({
-	// 					placeHolder: '',
-	// 					prompt: `Enter a value for the chain label: `
-	// 				});
-	// 				if (!chainLabel) return;
-	// 				const { src: xabis } = await rest.postContractsXabi(user, { src }, nodeOptions);
-	// 				const xabiKeys = Object.keys(xabis);
-	// 				const items = xabiKeys.map((x) => ({ label: x }))
-	// 				const quickPickOption = await vscode.window.showQuickPick(items, {
-	// 					placeHolder: 'Pick a contract to use as the chain\'s governance contract',
-	// 				});
-	// 				if (!quickPickOption) return;
-	// 				const contractName = quickPickOption ? quickPickOption.label : xabiKeys[0] || '';
-	// 				const govXabi = xabis[contractName] || {};
-	// 				let args = {}
-	// 				if (govXabi.constr) {
-	// 					const constr = govXabi.constr;
-	// 					const argNames = Object.keys(constr.args || {});
-	// 					for (let i = 0; i < argNames.length; i++) {
-	// 						const argInput = await vscode.window.showInputBox({
-	// 							placeHolder: '',
-	// 							prompt: `Enter a value for ${argNames[i]}: `
-	// 						});
-	// 						if (!argInput) return;
-	// 						args = { ...args, [argNames[i]]: argInput }
-	// 					}
-	// 				}
-	// 				const chainArgs = {
-	// 					label: chainLabel,
-	// 					src: srcMap,
-	// 					args,
-	// 					members: [{ "address": userAddress, "enode": "enode://abcd@1.2.3.4:30303" }],
-	// 					balances: [{ "address": userAddress, "balance": 100000000000000 }],
-	// 				}
-	// 				const contract = { name: contractName }
-	// 				const res = await rest.createChain(user, chainArgs, contract, nodeOptions);
-	// 				vscode.window.showInformationMessage(`${res}`);
-	// 			} catch (e) {
-	// 				vscode.window.showErrorMessage(`${e}`);
-	// 			}
-	// 		} else {
-	// 			vscode.window.showErrorMessage(`Please open a Solidity file to begin creating a private chain.`);
-	// 		}
-	// 	} else {
-	// 		vscode.window.showErrorMessage(`No active text editor found. Please open a Solidity file.`);
-	// 	}
-	// });
-
 
 	// Uploads a contract to the targetted node
 	vscode.commands.registerCommand('contracts.uploadContract', async (element) => {
@@ -300,7 +96,8 @@ export async function activate(context: vscode.ExtensionContext) {
 		const user = await getApplicationUser(nodeId);
 		const config = getConfig() || {}
 		const nodeOptions = { config, node: nodeId };
-		const { xabi } = await rest.getContractsContract(user, contractName, contractAddress, chainId, nodeOptions);
+		const val = await rest.getContractsContract(user, contractName, contractAddress, chainId, nodeOptions);
+		const { xabi } = val;
 		const func = ((xabi || {}).funcs || {})[variableName]
 		if (variableName && variableName !== 'constructor' && func) {
 			const argNames = Object.keys(func.args || {});
@@ -332,19 +129,17 @@ export async function activate(context: vscode.ExtensionContext) {
 	});
 
 
-	// Open the STRATO settings page
-	vscode.commands.registerCommand('project-management.settings', async () => {
-		// Open the user settings file
-		vscode.commands.executeCommand('workbench.action.openSettings', 'strato-vscode');
-	});
-
-
-	// Register the quick tools provider
-	vscode.window.registerTreeDataProvider('project-management', new ProjectActionProvider())
-
 	// Register the contracts provider for the sidebar
 	const contractsProvider = new ContractsProvider();
 	vscode.commands.registerCommand('contracts.refreshEntry', async () => contractsProvider.refresh());
+	vscode.commands.registerCommand('contracts.searchContract', async () => {
+		const argInput = await vscode.window.showInputBox({
+			placeHolder: 'ex: Certificate',
+			prompt: `Search for a contract by name`
+		});
+		if (!argInput) return;
+		contractsProvider.searchContracts(argInput);
+	})
 	vscode.window.registerTreeDataProvider('contracts', contractsProvider)
 
 	// Register the Cirrus provider
@@ -361,27 +156,29 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	// Register the nodes provider
 	const nodesProvider = new NodesProvider();
-	vscode.commands.registerCommand('nodes.refreshEntry', async () => {
-		let fp = context.workspaceState.get('strato-vscode.configPath')
-		console.debug(`refreshEntry/fp init: ${fp}`)
-		if (!fp) {
-			const options: vscode.OpenDialogOptions = {
-				title: 'Import configuration',
-				openLabel: 'Import configuration',
-				canSelectMany: false,
-				canSelectFiles: true,
-				canSelectFolders: false
-			};
-			const userSelect = await vscode.window.showOpenDialog(options) || '';
-			console.debug(`refreshEntry/userSelect: ${userSelect}`)
+	vscode.commands.registerCommand('nodes.addConfig', async () => {
+		let fp = vscode.workspace.getConfiguration().get('strato-vscode.configPath') || '';
+		const options: vscode.OpenDialogOptions = {
+			title: 'Import configuration',
+			openLabel: 'Import configuration',
+			canSelectMany: false,
+			canSelectFiles: true,
+			canSelectFolders: false
+		};
+		const userSelect = await vscode.window.showOpenDialog(options) || '';
+		console.debug(`refreshEntry/userSelect: ${userSelect}`)
 
-			fp = userSelect && userSelect[0] ? vscode.Uri.parse(userSelect[0].fsPath) : null
-			console.debug(`refreshEntry/fp: ${fp}`)
-			context.workspaceState.update('strato-vscode.configPath', fp)
-			console.debug(`refreshEntry/workspaceState: ${context.workspaceState.get('strato-vscode.configPath')} `)
-		}
+		fp = userSelect && userSelect[0] ? userSelect[0].fsPath : ''
+		await vscode.workspace.getConfiguration().update('strato-vscode.configPath', fp)
 		nodesProvider.refresh()
 	})
+	
+	// Open the STRATO settings page
+	vscode.commands.registerCommand('nodes.settings', async () => {
+		// Open the user settings file
+		vscode.commands.executeCommand('workbench.action.openSettings', 'strato-vscode');
+	});
+
 	vscode.window.registerTreeDataProvider('nodes', nodesProvider)
 
 	// Register clipboard copier
@@ -444,6 +241,3 @@ export function copyToClipboard(t: string) {
 	vscode.env.clipboard.writeText(t)
 		.then(() => { vscode.window.showInformationMessage('Copied to clipboard') })
 }
-
-
-// todo(moncayo): accepts a vscode terminal and sets it to the root workspace dir